+++ src/main/java/base/example/demo/dto/CustomerDto.java
@@ -5,10 +5,10 @@ import jakarta.persistence.Column;
 public class CustomerDto {
     private String name;
     private String email;
-    private int phone;
-    private int pass;
+    private Long phone;
+    private Long pass;
 
-    public CustomerDto(String name, String email, int phone, int pass) {
+    public CustomerDto(String name, String email, Long phone, Long pass) {
         this.name = name;
         this.email = email;
         this.phone = phone;
@@ -25,11 +25,11 @@ public class CustomerDto {
         this.email = email;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(Long phone) {
         this.phone = phone;
     }
 
-    public void setPass(int pass) {
+    public void setPass(Long pass) {
         this.pass = pass;
     }
 
@@ -41,11 +41,11 @@ public class CustomerDto {
         return email;
     }
 
-    public int getPhone() {
+    public Long getPhone() {
         return phone;
     }
 
-    public int getPass() {
+    public Long getPass() {
         return pass;
     }
 // сет гет в данном случае не нужны, тк объект передает данные в своих полях
